- Moved synchronization and search packages to new components module. - Added ServerLog. - Added Shopify data classed. - Added TradesmanLead.

+++ src/main/java/com/fixit/core/dao/sql/ReviewDao.java
@@ -4,5 +4,5 @@ import com.fixit.core.data.sql.Review;
 import com.fixit.core.data.sql.Review.ReviewPK;
 
 public interface ReviewDao extends SqlDao<Review, ReviewPK> {
-	Integer getCountForTradesman(String tradesmanId);
+	Long getCountForTradesman(String tradesmanId);
 }
